Add propTypes
