+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "apollo": "^0.2.2",
+    "apollo-cache-inmemory": "^1.1.9",
     "apollo-client": "^2.2.1",
     "apollo-client-preset": "^1.0.6",
     "apollo-link-http": "^1.3.2",
