+++ saleor/static/dashboard/js/components/app/category/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { Route, Switch } from 'react-router-dom';
 
-import { CategoryCreateForm, CategoryUpdateForm } from './form';
 import CategoryDetails from './details';
+import { CategoryCreateForm, CategoryUpdateForm } from './form';
 
 const Component = () => (
   <Switch>
@@ -28,8 +27,5 @@ const Component = () => (
     />
   </Switch>
 );
-Component.propTypes = {
-  match: PropTypes.object,
-};
 
 export default Component;
